Moved PlayerSkinConfiguration
