+++ proxy/src/main/java/net/md_5/bungee/UserConnection.java
@@ -37,7 +37,7 @@ import net.md_5.bungee.api.event.PermissionCheckEvent;
 import net.md_5.bungee.api.event.ServerConnectEvent;
 import net.md_5.bungee.api.score.Scoreboard;
 import net.md_5.bungee.chat.ComponentSerializer;
-import net.md_5.bungee.conf.PlayerSkinConfiguration;
+import net.md_5.bungee.PlayerSkinConfiguration;
 import net.md_5.bungee.connection.InitialHandler;
 import net.md_5.bungee.entitymap.EntityMap;
 import net.md_5.bungee.forge.ForgeClientHandler;
